backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
openpyxl
langchain
langgraph
pydantic
python-dotenv
